Time: 0 ms (100.00%), Space: 6.2 MB (45.46%) - LeetHub

+++ 1328-break-a-palindrome/1328-break-a-palindrome.cpp
@@ -1,17 +1,16 @@
 class Solution {
 public:
     string breakPalindrome(string p) {
-         int sz = p.size();
-         if (sz <= 1) return "";
+        int sz = p.size() ;
+        if (sz <= 1) return "";
         
-        for (size_t i=0; i < sz/2; i++) {
+        for (int i=0; i < sz/2; i++) {
             if (p[i] > 'a')  {
                 p[i] = 'a';
                 return p;
             }
         }
-        
-        p[sz-1] = 'b';
+        p[sz-1] = 'b' ;
         return p;
     }
 };
